requirements.txt
flask>=2.0.0
werkzeug==2.0.1
pandas>=1.3.0
numpy==1.21.6
deezer-python>=3.0.0
gunicorn==20.1.0
billboard.py>=7.0.0
spotipy>=2.19.0
requests>=2.26.0
